+++ includes/dependencies.php
@@ -28,7 +28,6 @@ Autoloader::alias('Hybrid\Acl', 'Orchestra\Acl');
 Autoloader::alias('Hybrid\Form', 'Orchestra\Form');
 Autoloader::alias('Hybrid\HTML', 'Orchestra\HTML');
 Autoloader::alias('Hybrid\Memory', 'Orchestra\Memory');
-Autoloader::alias('Hybrid\Response', 'Orchestra\Response');
 Autoloader::alias('Hybrid\Table', 'Orchestra\Table');
 
 /*